backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.29
Flask-Restx==1.3.0
PyYAML==6.0.2
PyJWT==2.8.0
Flask-Cors==5.0.0